Modify MediaPlayerLoggingId to be a int64_t
The player id is sometimes stored as an int64_t, sometimes as an
int32_t. This will now achieve consistency with the generated
SetPlayerId function that is called through the UKM generated code.

Bug: 421435479
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6781628
Commit-Queue: Vikram Pasupathy <vpasupathy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1491137}

+++ media/base/media_player_logging_id.h
@@ -14,7 +14,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace media {
 
 // alias the name for readability
-using MediaPlayerLoggingID = int32_t;
+using MediaPlayerLoggingID = int64_t;
 
 MEDIA_EXPORT MediaPlayerLoggingID GetNextMediaPlayerLoggingID();
